updates to navigation, naming conventions
