--- src/components/Login.css
@@ -1,3 +1,0 @@
-div {
-    background-color: gray;
-}
