--- src/components/NavBar.css
@@ -1,3 +1,0 @@
-.navBar {
-	background-color: lightskyblue;
-}
