+++ src/components/NavBar.jsx
@@ -1,24 +1,30 @@
-import React from 'react';
-import { NavLink } from 'react-router-dom';
-import './style.css';
+import React from "react";
+import { NavLink } from "react-router-dom";
+import "./style.css";
 
 export const NavBar = ({ currentUser }) => {
-	if (currentUser) {
-		return (
-			<nav className="navBar">
-				<NavLink to="/allproducts">All Products</NavLink>
-				<NavLink to='/login'>Login</NavLink>
-				<NavLink to='/register'>Register</NavLink>
-
-			</nav>
-		);
-	} else {
-		return (
-			<nav>
-				<NavLink to="/allproducts">All Products</NavLink>
-			</nav>
-		);
-	}
+    if (currentUser) {
+        return (
+            <nav className="nav-bar">
+                <NavLink to="/" className="nav-item">
+                    Home
+                </NavLink>
+                <NavLink to="/products" className="nav-item">
+                    Products
+                </NavLink>
+                <NavLink to="/login" className="nav-item">
+                    Login
+                </NavLink>
+                <NavLink to="/register" className="nav-item">
+                    Register
+                </NavLink>
+            </nav>
+        );
+    } else {
+        return (
+            <nav>
+                <NavLink to="/products">Products</NavLink>
+            </nav>
+        );
+    }
 };
-
-export default NavBar;
