+++ src/components/Product.jsx
@@ -1,24 +1,33 @@
-import React from 'react';
-import { Link } from 'react-router-dom';
-import './style.css';
+import React, { useState, useEffect } from "react";
+import { getProduct } from "../api";
+import { useParams } from "react-router-dom";
 
-export const Product = ({ product }) => {
-	const { id, name, description, price, imageURL, inStock, category } =
-		product;
+import "./style.css";
 
-	return (
-		<div id={`product${id}`} className="product-card">
-			<Link to={`/product/${id}`}>
-				<h3>{name}</h3>
-			</Link>
-			<img src={imageURL} alt="product" />
-			<h4>Product Information</h4>
-			<p>Category: {category}</p>
-			<p>description: {description}</p>
-			<p>Price: ${price}</p>
-			<p>In Stock: {inStock}</p>
-		</div>
-	);
+export const Product = (props) => {
+    const { productId } = useParams();
+    const [product, setProduct] = useState([]);
+
+    useEffect(() => {
+        const callback = async () => {
+            const prod = await getProduct(productId);
+            setProduct(prod);
+        };
+        callback();
+    }, [productId]);
+
+    const { id, name, description, price, imageURL, inStock, category } =
+        product;
+
+    return (
+        <div key={productId} id={`singleProduct${id}`} className="product-card">
+            <h3>{name}</h3>
+            <img src={imageURL} alt="product" />
+            <h4>Product Information</h4>
+            <p>Category: {category}</p>
+            <p>description: {description}</p>
+            <p>Price: ${price}</p>
+            <p>In Stock: {inStock}</p>
+        </div>
+    );
 };
-
-export default Product;
