--- src/components/Products.css
@@ -1,17 +1,0 @@
-.allProducts {
-	background-color: gray;
-	font-size: large;
-}
-
-.product-card {
-	border: 3px black solid;
-	height: 400px;
-	width: 250px;
-	margin: 5px;
-}
-
-img {
-	border-radius: 50%;
-	height: 50px;
-	width: 50px;
-}
